+++ AnalyseGenomeMFC/AnalyseGenomeMFC/Maladie.h
@@ -9,16 +9,16 @@ public:
 	Maladie();
 	~Maladie();
 
-	long get_id();
+	int get_id();
 	string get_nom();
 	string get_typeMaladie();
 
-	void set_id(long id_in);
+	void set_id(int id_in);
 	void set_nom(string name_in);
 	void set_typeMaladie(string name_in);
 
 private:
-	long id;
+	int id;
 	string nom, typeMaladie;
 };
