+++ AnalyseGenomeMFC/AnalyseGenomeMFC/Medecin.h
@@ -7,7 +7,7 @@ using namespace std;
 class Medecin
 {
 private:
-	long id;
+	int id;
 	string nom;
 	string prenom;
 	string adresse;
@@ -18,7 +18,7 @@ public:
 	Medecin();
 	~Medecin();
 
-	void set_id(long id_in);
+	void set_id(int id_in);
 	void set_nom(string nom_in);
 	void set_prenom(string prenom_in);
 	void set_adresse(string mail_in);
@@ -26,7 +26,7 @@ public:
 	void set_salaire(double sal_in);
 
 
-	long get_id();
+	int get_id();
 	string get_nom();
 	string get_prenom();
 	string get_adresse();
