+++ AnalyseGenomeMFC/AnalyseGenomeMFC/Serveur.h
@@ -8,18 +8,18 @@ class Serveur
 public:
 	Serveur();
 	~Serveur();
-	void set_id(long id_in);
+	void set_id(int id_in);
 	void set_IPServeur(string ip_in);
 	void set_description(string desc_in);
-	void set_idEntreprise(long entr_in);
+	void set_idEntreprise(int entr_in);
 
-	long get_id();
+	int get_id();
 	string get_IPServeur();
 	string get_description();
-	long get_idEntreprise();
+	int get_idEntreprise();
 private:
-	long id;
+	int id;
 	string IPServeur, description;
-	long idEntreprise;
+	int idEntreprise;
 };
